Complete cloud domain boundary extraction
