+++ apps/desktop/src/main/cloud/control-plane-domains/channels.ts
@@ -25,4 +25,8 @@ export type ChannelControlPlaneStore = Pick<ControlPlaneStore,
   | 'listChannelDeliveries'
   | 'claimNextChannelDelivery'
   | 'ackChannelDelivery'
+  | 'getSession'
+  | 'getSessionProjection'
+  | 'enqueueSessionCommand'
+  | 'recordAuditEvent'
 >
