+++ packages/cloud-client/src/domains/capabilities.js
@@ -1,0 +1,1 @@
+export * from './capabilities.ts'
