+++ packages/cloud-client/src/domains/config.js
@@ -1,0 +1,1 @@
+export * from './config.ts'
