+++ packages/cloud-client/src/domains/transport.ts
@@ -1,17 +1,32 @@
 export type {
   CloudTransportAdapter,
   CloudTransportAdapterOptions,
-  CloudTransportErrorKind,
-  CloudTransportErrorOptions,
   CloudTransportEventSource,
   CloudTransportFetch,
   CloudTransportResponse,
   CloudTransportSessionEvent,
   CloudTransportSubscription,
   CloudTransportWorkspaceEvent,
-} from '../adapter.js'
+} from '../contracts.js'
 
 export {
   CloudTransportError,
   isCloudTransportError,
-} from '../adapter.js'
+} from '../errors.js'
+
+export type {
+  CloudTransportErrorKind,
+  CloudTransportErrorOptions,
+} from '../errors.js'
+
+export {
+  createCloudTransportEventClient,
+  sessionEventsUrl,
+  subscribeCloudEvents,
+  workspaceEventsUrl,
+} from '../domain-clients/transport.js'
+
+export type {
+  CloudTransportEventClient,
+  CloudTransportSseContext,
+} from '../domain-clients/transport.js'
